refactor pools state
